backend/requirements-dev.txt
# Include production requirements
-r requirements.txt

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Type Checking
mypy==1.7.1
django-stubs==4.2.7
djangorestframework-stubs==3.14.5

# Linting and Formatting
ruff==0.1.6

# Development Tools
ipython==8.18.1
django-extensions==3.2.3
